fix recycler view errors - make radio buttons functional - change add pastoral team and images to dialog layouts
