+++ companion/app/src/main/java/com/kingstek/companion/ui/parish/updates/UpdateParishInfoFragment.kt
@@ -12,19 +12,18 @@ import android.text.Editable
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.AdapterView
-import android.widget.ArrayAdapter
-import android.widget.Spinner
-import android.widget.Toast
+import android.widget.*
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.app.ActivityCompat
 import androidx.core.content.ContextCompat.checkSelfPermission
+import androidx.core.graphics.green
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.kingstek.companion.R
 import com.kingstek.companion.databinding.FragmentUpdateParishInfoBinding
 import com.kingstek.companion.models.parish.ImageModel
 import com.kingstek.companion.models.parish.Parish
@@ -38,7 +37,6 @@ class UpdateParishInfoFragment : Fragment() {
     private var _binding: FragmentUpdateParishInfoBinding? = null
     private val binding get() =  _binding!!
     private lateinit var viewModel: UpdateParishInfoViewModel
-    private lateinit var imgURI: Uri
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         // Inflate the layout for this fragment
@@ -63,40 +61,51 @@ class UpdateParishInfoFragment : Fragment() {
          */
         //todo only set parish list spinner when selecting update parish info
         //todo set parish coming from as default selected spinner value
-        val parishSpinner: Spinner = binding.spParishListSpinner
 
-        binding.etParishName.setOnClickListener{
-            binding.spParishListSpinner.performClick()
-        }
+        viewModel.isUpdateOrNew.postValue(true)
 
-        binding.spParishListSpinner.onItemSelectedListener = object : AdapterView.OnItemSelectedListener {
-            override fun onItemSelected(
-                parent: AdapterView<*>,
-                view: View,
-                position: Int,
-                id: Long
-            ) {
-                // Get the selected item from the adapter
-                val selectedItem = parent.getItemAtPosition(position).toString()
-                val editable = Editable.Factory.getInstance().newEditable(selectedItem)
+        viewModel.isUpdateOrNew.observe(viewLifecycleOwner, Observer {
+            if (it == true) {
+                val parishSpinner: Spinner = binding.spParishListSpinner
+
+                binding.etParishName.setOnClickListener{
+                    binding.spParishListSpinner.performClick()
+                }
+
+                binding.spParishListSpinner.onItemSelectedListener = object : AdapterView.OnItemSelectedListener {
+                    override fun onItemSelected(
+                        parent: AdapterView<*>,
+                        view: View,
+                        position: Int,
+                        id: Long
+                    ) {
+                        // Get the selected item from the adapter
+                        val selectedItem = parent.getItemAtPosition(position).toString()
+                        val editable = Editable.Factory.getInstance().newEditable(selectedItem)
+                        binding.etParishName.text = editable
+                    }
+
+                    override fun onNothingSelected(parent: AdapterView<*>?) {
+                        // Do nothing
+                        // to if nothing selected display error
+                    }
+                }
+
+                //spinner adapter
+                val dataAdapter: ArrayAdapter<String> = ArrayAdapter(
+                    requireContext(),
+                    android.R.layout.simple_spinner_item,
+                    viewModel.parishListSpinner.value!!
+                )
+
+                dataAdapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item)
+                parishSpinner.adapter = dataAdapter
+            } else {
+                val editable = Editable.Factory.getInstance().newEditable("")
                 binding.etParishName.text = editable
             }
+        } )
 
-            override fun onNothingSelected(parent: AdapterView<*>?) {
-                // Do nothing
-                // to if nothing selected display error
-            }
-        }
-
-        //spinner adapter
-        val dataAdapter: ArrayAdapter<String> = ArrayAdapter(
-            requireContext(),
-            android.R.layout.simple_spinner_item,
-            viewModel.parishListSpinner.value!!
-        )
-
-        dataAdapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item)
-        parishSpinner.adapter = dataAdapter
 
 
         /**
@@ -191,44 +200,13 @@ class UpdateParishInfoFragment : Fragment() {
 
         /**
          * add edit views to pastoral team list that will be passed in when pushing parish details
-         * toggles visibility of edit views for adding team details
+         * toggles visibility of edit views for adding pastoral team details
          */
         binding.tvAddBtn.setOnClickListener {
 
-            binding.ltPastoralTeam.visibility = View.VISIBLE
+            val dialogFragment = PastoralTeamDialog(viewModel, pastoralAdapter)
+            dialogFragment.show(childFragmentManager, "pastoralDialog")
 
-            if (!validatePastoralTeam()) {
-                return@setOnClickListener
-            }
-
-            val pastoralTeam = PastoralTeam(
-                binding.etPosition.text.toString(),
-                binding.etTitle.text.toString(),
-                binding.etName.text.toString(),
-                binding.etPhoneNumber.text.toString()
-            )
-
-            viewModel.addPastoralTeam(pastoralTeam)
-
-            // is this necessary already being done in adapter
-            pastoralAdapter.notifyDataSetChanged()
-
-            val emptyString = ""
-            val editable = Editable.Factory.getInstance().newEditable(emptyString)
-
-            binding.etPosition.text = editable
-            binding.etTitle.text = editable
-            binding.etName.text = editable
-            binding.etPhoneNumber.text = editable
-        }
-
-        /**
-         * Toggle pastoral team add and done buttons
-         * toggles visibility of edit views for adding team details
-         */
-
-        binding.tvDoneBtn.setOnClickListener {
-            binding.ltPastoralTeam.visibility = View.GONE
         }
 
         /**
@@ -250,39 +228,42 @@ class UpdateParishInfoFragment : Fragment() {
 
         binding.tvImagesAddBtn.setOnClickListener {
 
-            binding.ltParishImages.visibility = View.VISIBLE
+            val dialogFragment = ParishImageDialog(viewModel, parishImageAdapter)
+            dialogFragment.show(childFragmentManager, "parishImageDialog")
 
-            if (binding.etImageTitle.text?.isEmpty()!!) {
-                //todo add a check for empty images too
-                //todo add a check for empty lists
-                binding.tilImageTitle.error = "Enter Image tile and select Image"
-                return@setOnClickListener
-            }
-
-            val parishImages = ImageModel(
-                imgURI,
-                binding.etImageTitle.text.toString(),
-            )
-
-            viewModel.addParishImageList(parishImages)
-
-            // is this necessary already being done in adapter
-            pastoralAdapter.notifyDataSetChanged()
-
-            val emptyString = ""
-            val editable = Editable.Factory.getInstance().newEditable(emptyString)
-
-            binding.etImageTitle.text = editable
-            binding.etImageTitle.error = null
+//            binding.ltParishImages.visibility = View.VISIBLE
+//
+//            if (binding.etImageTitle.text?.isEmpty()!!) {
+//                //todo add a check for empty images too
+//                //todo add a check for empty lists
+//                binding.tilImageTitle.error = "Enter Image tile and select Image"
+//                return@setOnClickListener
+//            }
+//
+//            val parishImages = ImageModel(
+//                imgURI,
+//                binding.etImageTitle.text.toString(),
+//            )
+//
+//            viewModel.addParishImageList(parishImages)
+//
+//            // is this necessary already being done in adapter
+//            pastoralAdapter.notifyDataSetChanged()
+//
+//            val emptyString = ""
+//            val editable = Editable.Factory.getInstance().newEditable(emptyString)
+//
+//            binding.etImageTitle.text = editable
+//            binding.etImageTitle.error = null
         }
 
-        binding.tvImagesDoneBtn.setOnClickListener {
-            binding.ltParishImages.visibility = View.GONE
-        }
+//        binding.tvImagesDoneBtn.setOnClickListener {
+//            binding.ltParishImages.visibility = View.GONE
+//        }
 
-        binding.ivSelectedImage.setOnClickListener {
-            pickImage()
-        }
+//        binding.ivSelectedImage.setOnClickListener {
+//            pickImage()
+//        }
 
         //todo use alert dialogs to add images and pastoral team
         binding.btnPublish.setOnClickListener {
@@ -301,6 +282,14 @@ class UpdateParishInfoFragment : Fragment() {
             viewModel.uploadParishInfo(requireContext())
 
 //            setValues()
+        }
+
+        binding.rbUpdateBtn.setOnClickListener {
+            onRadioButtonClicked(it)
+        }
+
+        binding.rbNewAdditionBtn.setOnClickListener {
+            onRadioButtonClicked(it)
         }
 
         return root
@@ -371,58 +360,64 @@ class UpdateParishInfoFragment : Fragment() {
     /**
      * Pastoral team setup
      */
-    fun validatePastoralTeam(): Boolean {
+//    fun validatePastoralTeam(): Boolean {
+//
+//        if(binding.etTitle.text?.isEmpty()!!) {
+//            binding.tilTitle.error = "Field must not be empty"
+//            return false
+//        }
+//
+//        binding.tilTitle.error = null
+//
+//        if(binding.etName.text?.isEmpty()!!) {
+//            binding.tilName.error = "Field must not be empty"
+//            return false
+//        }
+//
+//        binding.tilName.error = null
+//
+//        if(binding.etPosition.text?.isEmpty()!!) {
+//            binding.tilPosition.error = "Field must not be empty"
+//            return false
+//        }
+//
+//        binding.tilPosition.error = null
+//
+//        if(binding.etPhoneNumber.text?.isEmpty()!!) {
+//            binding.tilPhonenumber.error = "Field must not be empty"
+//            return false
+//        }
+//
+//        binding.tilPhonenumber.error = null
+//
+//        return true
+//    }
 
-        if(binding.etTitle.text?.isEmpty()!!) {
-            binding.tilTitle.error = "Field must not be empty"
-            return false
+//    private fun pickImage() {
+//        val intent = Intent()
+//        intent.type = "image/*"
+//        intent.action = (Intent.ACTION_GET_CONTENT)
+//
+//        pickImageLauncher.launch(intent)
+//    }
+
+    /**
+     * Toogle update and new addition radio buttons
+     */
+    fun onRadioButtonClicked(view: View) {
+        val checked = (view as RadioButton).isChecked
+        when (view.getId()) {
+            R.id.rb_update_btn -> {
+                if (checked) viewModel.isUpdateOrNew.postValue(true)
+            }
+
+            R.id.rb_new_addition_btn -> {
+                if (checked) viewModel.isUpdateOrNew.postValue(false)
+            }
         }
-
-        binding.tilTitle.error = null
-
-        if(binding.etName.text?.isEmpty()!!) {
-            binding.tilName.error = "Field must not be empty"
-            return false
-        }
-
-        binding.tilName.error = null
-
-        if(binding.etPosition.text?.isEmpty()!!) {
-            binding.tilPosition.error = "Field must not be empty"
-            return false
-        }
-
-        binding.tilPosition.error = null
-
-        if(binding.etPhoneNumber.text?.isEmpty()!!) {
-            binding.tilPhonenumber.error = "Field must not be empty"
-            return false
-        }
-
-        binding.tilPhonenumber.error = null
-
-        return true
     }
 
-    private fun pickImage() {
-        val intent = Intent()
-        intent.type = "image/*"
-        intent.action = (Intent.ACTION_GET_CONTENT)
 
-        pickImageLauncher.launch(intent)
-    }
-
-    private var pickImageLauncher = registerForActivityResult(ActivityResultContracts.StartActivityForResult()) {
-        if (it.resultCode == Activity.RESULT_OK) {
-            val data: Intent? = it.data
-
-            val uri: Uri? = data?.data
-
-            imgURI = uri!!
-            val bitmap = MediaStore.Images.Media.getBitmap(requireContext().contentResolver, uri)
-            binding.ivSelectedImage.setImageBitmap(bitmap)
-        }
-    }
 
     private fun checkPermissionForReadExtertalStorage(): Boolean {
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
